Get corresponding raster maps to feature

Specify sf Object an some additional parameters to get corresponding raster maps

get_raster(features, scale_level, x_add = 0, y_add = 0,
  per_feature = F, method = "centroid", turn_greyscale = F,
  name = "", fdir = NULL, limit = Inf)

Arguments

features
The sf object to derive the raster data from

scale_level
The scale at which to get raster data (x in 1:1'000x). Usually one of the following values (depending on the available rasters): 10,25,100,500,1000

x_add, y_add
Depending on method, x and y will be added to the centeroid or to the bounding box

per_feature
A TRUE/FALSE value specifying if one raster map should be returend for the entire sf object or if one map should be returned per feature

method
A character string specifying the method with which the extent should be calculated. centroid calculates the centroid of the object(s), bbox calculates the bounding box of the object.

turn_greyscale
Should the output rastermaps be turned into greyscale?

name
If muliple, different maps are available with overlapping extents, name can be used to differentiate between different maptypes. Default is an empty sting.

fdir
By default, the fdir is retrieved from the package Environment (named packageEnv). Override this with fdir =

limit
For testing puposes only: Limits the number of rasters returned per object. Defaults to Inf