Show extents of Available Raste Files

This function visualizes the extents of all available rasters. The plots are faceted by scale and the area is colorized by resolution. The extents have an alpha value so that overlapping extents can be detected.

show_extents(method = "ggplot2", fdir = NULL)

Arguments

method
The method with which to visualize the data: ggplot2 or tmap Wrap the output in ggplotly() to get an interactive ggplot output or set tmap_mode to "view" go get an interactive tmap.

filedirectory
The file directory aquired by init_fdir(). If left as NULL, the fdir from the package Environment is taken.

Details

This function requires ggplot2 and / or tmap, depending on the output type defined in "method".